Specialize ResourceHandler as MappedByteBufferHandler
