+++ src/main/java/com/metamx/common/FileUtils.java
@@ -138,45 +138,24 @@ public class FileUtils
    *
    * <p>This only works for files <= {@link Integer#MAX_VALUE} bytes.
    *
-   * <p>Similar to {@link Files#map(File)}, but returns {@link ResourceHandler}, that makes it easier to unmap the
-   * buffer within try-with-resources pattern:
+   * <p>Similar to {@link Files#map(File)}, but returns {@link MappedByteBufferHandler}, that makes it easier to unmap
+   * the buffer within try-with-resources pattern:
    * <pre>{@code
-   * try (ResourceHandler<MappedByteBuffer> fileMappingHandler = FileUtils.map(file)) {
+   * try (MappedByteBufferHandler fileMappingHandler = FileUtils.map(file)) {
    *   ByteBuffer fileMapping = fileMappingHandler.get();
    *   // use mapped buffer
    * }}</pre>
    *
    * @param file the file to map
-   * @return a {@link ResourceHandler}, wrapping a read-only buffer reflecting {@code file}
+   * @return a {@link MappedByteBufferHandler}, wrapping a read-only buffer reflecting {@code file}
    * @throws FileNotFoundException if the {@code file} does not exist
    * @throws IOException if an I/O error occurs
    *
    * @see FileChannel#map(FileChannel.MapMode, long, long)
    */
-  public static ResourceHandler<MappedByteBuffer> map(File file) throws IOException
+  public static MappedByteBufferHandler map(File file) throws IOException
   {
     MappedByteBuffer mappedByteBuffer = Files.map(file);
     return new MappedByteBufferHandler(mappedByteBuffer);
-  }
-
-  private static class MappedByteBufferHandler implements ResourceHandler<MappedByteBuffer> {
-    private final MappedByteBuffer mappedByteBuffer;
-
-    private MappedByteBufferHandler(MappedByteBuffer mappedByteBuffer)
-    {
-      this.mappedByteBuffer = mappedByteBuffer;
-    }
-
-    @Override
-    public MappedByteBuffer get()
-    {
-      return mappedByteBuffer;
-    }
-
-    @Override
-    public void close()
-    {
-      ByteBufferUtils.unmap(mappedByteBuffer);
-    }
   }
 }
